+++ plugins/source/azure/resources/services/compute/instance_views.go
@@ -24,7 +24,7 @@ func instanceViews() *schema.Table {
 			{
 				Name:     "compute_virtual_machine_id",
 				Type:     schema.TypeUUID,
-				Resolver: schema.ParentIDResolver,
+				Resolver: schema.ParentResourceFieldResolver("id"),
 			},
 			{
 				Name:     "platform_update_domain",