+++ plugins/source/azure/resources/services/cosmosdb/mongo_db_databases.go
@@ -24,7 +24,7 @@ func mongoDBDatabases() *schema.Table {
 			{
 				Name:     "cosmosdb_account_id",
 				Type:     schema.TypeUUID,
-				Resolver: schema.ParentIDResolver,
+				Resolver: schema.ParentResourceFieldResolver("id"),
 			},
 			{
 				Name:     "resource",