+++ plugins/source/azure/resources/services/sql/server_blob_auditing_policies.go
@@ -24,7 +24,7 @@ func serverBlobAuditingPolicies() *schema.Table {
 			{
 				Name:     "sql_server_id",
 				Type:     schema.TypeUUID,
-				Resolver: schema.ParentIDResolver,
+				Resolver: schema.ParentResourceFieldResolver("id"),
 			},
 			{
 				Name:     "state",