+++ plugins/source/azure/resources/services/sql/transparent_data_encryptions.go
@@ -24,7 +24,7 @@ func transparentDataEncryptions() *schema.Table {
 			{
 				Name:     "sql_database_id",
 				Type:     schema.TypeUUID,
-				Resolver: schema.ParentIDResolver,
+				Resolver: schema.ParentResourceFieldResolver("id"),
 			},
 			{
 				Name:     "location",