+++ plugins/source/k8s/client/client.go
@@ -1,13 +1,15 @@
 package client
 
 import (
+	"context"
 	"fmt"
 	"strings"
 
-	"github.com/cloudquery/cq-provider-sdk/provider/diag"
-	"github.com/cloudquery/cq-provider-sdk/provider/schema"
-	"github.com/hashicorp/go-hclog"
+	"github.com/cloudquery/plugin-sdk/schema"
+	"github.com/cloudquery/plugin-sdk/specs"
+	"github.com/rs/zerolog"
 	"k8s.io/client-go/kubernetes"
+
 	// import all k8s auth options
 	_ "k8s.io/client-go/plugin/pkg/client/auth"
 	"k8s.io/client-go/tools/clientcmd"
@@ -15,34 +17,37 @@ import (
 )
 
 type Client struct {
-	Log      hclog.Logger
+	logger zerolog.Logger
+	// map context_name -> Services struct
 	services map[string]Services
-	kConfig  api.Config
-	config   *Config
+	spec     *Spec
 	contexts []string
 	paths    map[string]struct{}
 
 	Context string
 }
 
-func (c *Client) Logger() hclog.Logger {
-	return c.Log
+func (c *Client) Logger() *zerolog.Logger {
+	return &c.logger
 }
 
 func (c *Client) Services() Services {
 	return c.services[c.Context]
 }
 
-func (c Client) WithContext(context string) *Client {
+// Don't confuse `k8sContext` with `context.ctx`! k8s-context is a k8s-term that refers to a k8s cluster.
+func (c Client) WithContext(k8sContext string) *Client {
 	return &Client{
-		Log:      c.Log.With("context", context),
+		logger:   c.logger.With().Str("context", k8sContext).Logger(),
 		services: c.services,
-		kConfig:  c.kConfig,
-		config:   c.config,
-		Context:  context,
+		spec:     c.spec,
+		contexts: c.contexts,
+		paths:    c.paths,
+		Context:  k8sContext,
 	}
 }
 
+// Used for testing
 func (c *Client) SetServices(s map[string]Services) {
 	c.services = s
 	contexts := make([]string, 0, len(s))
@@ -52,67 +57,70 @@ func (c *Client) SetServices(s map[string]Services) {
 	c.contexts = contexts
 }
 
-func Configure(logger hclog.Logger, config interface{}) (schema.ClientMeta, diag.Diagnostics) {
+func Configure(ctx context.Context, logger zerolog.Logger, s specs.Source) (schema.ClientMeta, error) {
+	var k8sSpec Spec
+
+	if err := s.UnmarshalSpec(&k8sSpec); err != nil {
+		return nil, fmt.Errorf("failed to unmarshal k8s spec: %w", err)
+	}
+
 	kubeConfig := clientcmd.NewNonInteractiveDeferredLoadingClientConfig(
 		clientcmd.NewDefaultClientConfigLoadingRules(),
 		&clientcmd.ConfigOverrides{},
 	)
-	kCfg, err := kubeConfig.RawConfig()
+	rawKubeConfig, err := kubeConfig.RawConfig()
 	if err != nil {
-		return nil, diag.FromError(err, diag.USER)
+		return nil, err
 	}
 
-	cfg := config.(*Config)
-
 	var contexts []string
-	switch len(cfg.Contexts) {
+	switch len(k8sSpec.Contexts) {
 	case 0:
-		logger.Debug("no context set in configuration using current default defined context", "context", kCfg.CurrentContext)
-		contexts = []string{kCfg.CurrentContext}
+		logger.Debug().Str("context", rawKubeConfig.CurrentContext).Msg("no context set in configuration using current default defined context")
+		contexts = []string{rawKubeConfig.CurrentContext}
 	case 1:
-		if cfg.Contexts[0] == "*" {
-			logger.Debug("loading all available configuration")
-			for cName := range kCfg.Contexts {
+		if k8sSpec.Contexts[0] == "*" {
+			logger.Debug().Msg("loading all available configuration")
+			for cName := range rawKubeConfig.Contexts {
 				contexts = append(contexts, cName)
 			}
 		} else {
-			if _, ok := kCfg.Contexts[cfg.Contexts[0]]; !ok {
-				return nil, diag.FromError(fmt.Errorf("context %q doesn't exist in kube configuration", cfg.Contexts[0]), diag.USER)
+			if _, ok := rawKubeConfig.Contexts[k8sSpec.Contexts[0]]; !ok {
+				return nil, fmt.Errorf("context %q doesn't exist in kube configuration", k8sSpec.Contexts[0])
 			}
-			contexts = []string{cfg.Contexts[0]}
+			contexts = []string{k8sSpec.Contexts[0]}
 		}
 	default:
-		for _, cName := range cfg.Contexts {
-			if _, ok := kCfg.Contexts[cName]; !ok {
-				return nil, diag.FromError(fmt.Errorf("context %q doesn't exist in kube configuration", cName), diag.USER)
+		for _, cName := range k8sSpec.Contexts {
+			if _, ok := rawKubeConfig.Contexts[cName]; !ok {
+				return nil, fmt.Errorf("context %q doesn't exist in kube configuration", cName)
 			}
 			contexts = append(contexts, cName)
 		}
 	}
 
 	if len(contexts) == 0 {
-		return nil, diag.FromError(fmt.Errorf("could not find any context"), diag.USER, diag.WithDetails("Try to add context, https://kubernetes.io/docs/reference/kubectl/cheatsheet/#kubectl-context-and-configuration"))
+		return nil, fmt.Errorf("could not find any context. Try to add context, https://kubernetes.io/docs/reference/kubectl/cheatsheet/#kubectl-context-and-configuration")
 	}
 
 	c := Client{
-		Log:      logger,
+		logger:   logger,
 		services: make(map[string]Services),
-		kConfig:  kCfg,
-		config:   cfg,
+		spec:     &k8sSpec,
 		contexts: contexts,
 		Context:  contexts[0],
 		paths:    make(map[string]struct{}),
 	}
 
 	for _, ctxName := range contexts {
-		logger.Info("creating k8s client for context", "context", ctxName)
-		kClient, err := buildKubeClient(kCfg, ctxName)
+		logger.Info().Str("context", ctxName).Msg("creating k8s client for context")
+		kClient, err := buildKubeClient(rawKubeConfig, ctxName)
 		if err != nil {
-			return nil, diag.FromError(fmt.Errorf("failed to build k8s client for context %q: %w", ctxName, err), diag.INTERNAL)
+			return nil, fmt.Errorf("failed to build k8s client for context %q: %w", ctxName, err)
 		}
 		c.paths, err = getAPIsMap(kClient)
 		if err != nil {
-			c.Logger().Warn("Failed to get OpenAPI schema. It might be not supported in the current version of Kubernetes. OpenAPI has been supported since Kubernetes 1.4", "err", err)
+			logger.Warn().Err(err).Msg("Failed to get OpenAPI schema. It might be not supported in the current version of Kubernetes. OpenAPI has been supported since Kubernetes 1.4")
 		}
 		c.services[ctxName] = initServices(kClient)
 	}