+++ plugins/destination/postgresql/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## [0.3.0](https://github.com/cloudquery/cloudquery/compare/plugins-destination-postgresql-v0.2.1...plugins-destination-postgresql-v0.3.0) (2022-09-28)
+
+
+### Features
+
+* Move to standalone postgresql plugin ([#2074](https://github.com/cloudquery/cloudquery/issues/2074)) ([a0de6d3](https://github.com/cloudquery/cloudquery/commit/a0de6d3dfc0f43aad9b465c469b92a96121db0a1))
+* Update log ([#2051](https://github.com/cloudquery/cloudquery/issues/2051)) ([6efdf1c](https://github.com/cloudquery/cloudquery/commit/6efdf1c44434f7de13ab25a1f48d47a877c26a79))
+
+
+### Bug Fixes
+
+* **deps:** Update module github.com/cloudquery/plugin-sdk to v0.10.2 ([#2048](https://github.com/cloudquery/cloudquery/issues/2048)) ([e407991](https://github.com/cloudquery/cloudquery/commit/e4079914772d8191639b9935aa5970b8e27b082f))
+
 ## [0.2.1](https://github.com/cloudquery/cloudquery/compare/plugins-destination-postgresql-v0.2.0...plugins-destination-postgresql-v0.2.1) (2022-09-28)